+++ priv/repo/migrations/20251113033238_create_jobs.exs
@@ -6,8 +6,8 @@ defmodule App.Repo.Migrations.CreateJobs do
       add :title, :string
       add :payment_type, :string
       add :number_of_payouts, :integer
-      add :payout_amount, :float
-      add :hourly_rate, :float
+      add :payout_amount, :decimal
+      add :hourly_rate, :decimal
       add :company_id, references(:companies, on_delete: :nothing)
       add :user_id, references(:users, type: :id, on_delete: :delete_all)
